+++ engine/crates/integration-tests/src/mocks/graphql.rs
@@ -1,12 +1,13 @@
 //! A mock GraphQL server for testing the GraphQL connector
 
-use std::{net::TcpListener, time::Duration};
+use std::{net::TcpListener, sync::Arc, time::Duration};
 
-use async_graphql::{
-    EmptyMutation, EmptySubscription, InputObject, Interface, Object, Schema, SimpleObject, Union, ID,
-};
 use async_graphql_axum::{GraphQLRequest, GraphQLResponse};
-use axum::{http::HeaderMap, routing::post, Router};
+use axum::{extract::State, http::HeaderMap, routing::post, Router};
+
+mod fake_github;
+
+pub use fake_github::FakeGithubSchema;
 
 pub struct MockGraphQlServer {
     shutdown: Option<tokio::sync::oneshot::Sender<()>>,
@@ -22,8 +23,15 @@ impl Drop for MockGraphQlServer {
 }
 
 impl MockGraphQlServer {
-    pub async fn new() -> MockGraphQlServer {
-        let app = Router::new().route("/", post(graphql_handler));
+    pub async fn new(schema: impl Schema + 'static) -> MockGraphQlServer {
+        let state = AppState {
+            schema: Arc::new(schema),
+        };
+        Self::new_impl(state).await
+    }
+
+    async fn new_impl(state: AppState) -> Self {
+        let app = Router::new().route("/", post(graphql_handler)).with_state(state);
 
         let socket = TcpListener::bind("127.0.0.1:0").unwrap();
         let port = socket.local_addr().unwrap().port();
@@ -33,7 +41,7 @@ impl MockGraphQlServer {
         tokio::spawn(async move {
             axum::Server::from_tcp(socket)
                 .unwrap()
-                .serve(app.into_make_service())
+                .serve(app.with_state(()).into_make_service())
                 .with_graceful_shutdown(async move {
                     shutdown_rx.await.ok();
                 })
@@ -55,7 +63,7 @@ impl MockGraphQlServer {
     }
 }
 
-async fn graphql_handler(headers: HeaderMap, req: GraphQLRequest) -> GraphQLResponse {
+async fn graphql_handler(State(state): State<AppState>, headers: HeaderMap, req: GraphQLRequest) -> GraphQLResponse {
     let headers = headers
         .into_iter()
         .map(|(name, value)| {
@@ -65,159 +73,36 @@ async fn graphql_handler(headers: HeaderMap, req: GraphQLRequest) -> GraphQLResp
             )
         })
         .collect();
-    let schema = Schema::build(Query { headers }, EmptyMutation, EmptySubscription).finish();
+    // let schema = Schema::build(Query { headers }, EmptyMutation, EmptySubscription).finish();
 
-    schema.execute(req.into_inner()).await.into()
+    state.schema.execute(headers, req.into_inner()).await.into()
 }
 
-struct Query {
-    headers: Vec<(String, String)>,
+#[derive(Clone)]
+struct AppState {
+    schema: Arc<dyn Schema>,
 }
 
-#[Object]
-impl Query {
-    // A top level scalar field for testing
-    async fn server_version(&self) -> &str {
-        "1"
+/// Creating a trait for schema so we can use it as a trait object and avoid
+/// making everything generic over Query, Mutation & Subscription params
+#[async_trait::async_trait]
+pub trait Schema: Send + Sync {
+    async fn execute(&self, headers: Vec<(String, String)>, request: async_graphql::Request)
+        -> async_graphql::Response;
+}
+
+#[async_trait::async_trait]
+impl<Q, M, S> Schema for async_graphql::Schema<Q, M, S>
+where
+    Q: async_graphql::ObjectType + 'static,
+    M: async_graphql::ObjectType + 'static,
+    S: async_graphql::SubscriptionType + 'static,
+{
+    async fn execute(
+        &self,
+        _headers: Vec<(String, String)>,
+        request: async_graphql::Request,
+    ) -> async_graphql::Response {
+        async_graphql::Schema::execute(self, request).await
     }
-
-    async fn pull_requests_and_issues(&self, _filter: PullRequestsAndIssuesFilters) -> Vec<PullRequestOrIssue> {
-        // This doesn't actually filter anything because I don't need that for my test.
-        vec![
-            PullRequestOrIssue::PullRequest(PullRequest {
-                title: "Creating the thing".into(),
-                checks: vec!["Success!".into()],
-                author: UserOrBot::User(User {
-                    name: "Jim".into(),
-                    email: "jim@example.com".into(),
-                }),
-            }),
-            PullRequestOrIssue::PullRequest(PullRequest {
-                title: "Some bot PR".into(),
-                checks: vec!["Success!".into()],
-                author: UserOrBot::Bot(Bot { id: "123".into() }),
-            }),
-            PullRequestOrIssue::Issue(Issue {
-                title: "Everythings fucked".into(),
-                author: UserOrBot::User(User {
-                    name: "The Pessimist".into(),
-                    email: "pessimist@example.com".into(),
-                }),
-            }),
-        ]
-    }
-
-    #[allow(unused_variables)]
-    async fn bot_pull_requests(&self, bots: Vec<Option<Vec<BotInput>>>) -> Vec<PullRequest> {
-        vec![
-            PullRequest {
-                title: "Creating the thing".into(),
-                checks: vec!["Success!".into()],
-                author: UserOrBot::User(User {
-                    name: "Jim".into(),
-                    email: "jim@example.com".into(),
-                }),
-            },
-            PullRequest {
-                title: "Some bot PR".into(),
-                checks: vec!["Success!".into()],
-                author: UserOrBot::Bot(Bot { id: "123".into() }),
-            },
-        ]
-    }
-
-    async fn pull_request_or_issue(&self, id: ID) -> Option<PullRequestOrIssue> {
-        if id == "1" {
-            return Some(PullRequestOrIssue::PullRequest(PullRequest {
-                title: "Creating the thing".into(),
-                checks: vec!["Success!".into()],
-                author: UserOrBot::User(User {
-                    name: "Jim".into(),
-                    email: "jim@example.com".into(),
-                }),
-            }));
-        } else if id == "2" {
-            return Some(PullRequestOrIssue::PullRequest(PullRequest {
-                title: "Some bot PR".into(),
-                checks: vec!["Success!".into()],
-                author: UserOrBot::Bot(Bot { id: "123".into() }),
-            }));
-        } else if id == "3" {
-            return Some(PullRequestOrIssue::Issue(Issue {
-                title: "Everythings fucked".into(),
-                author: UserOrBot::User(User {
-                    name: "The Pessimist".into(),
-                    email: "pessimist@example.com".into(),
-                }),
-            }));
-        }
-        None
-    }
-
-    async fn headers(&self) -> Vec<Header> {
-        self.headers
-            .clone()
-            .into_iter()
-            .map(|(name, value)| Header { name, value })
-            .collect()
-    }
-}
-
-#[derive(SimpleObject)]
-struct Header {
-    name: String,
-    value: String,
-}
-
-#[derive(SimpleObject)]
-struct PullRequest {
-    title: String,
-    checks: Vec<String>,
-    author: UserOrBot,
-}
-
-#[derive(SimpleObject)]
-struct Issue {
-    title: String,
-    author: UserOrBot,
-}
-
-#[derive(Interface)]
-#[graphql(field(name = "title", ty = "String"), field(name = "author", ty = "UserOrBot"))]
-enum PullRequestOrIssue {
-    PullRequest(PullRequest),
-    Issue(Issue),
-}
-
-#[derive(Union, Clone)]
-enum UserOrBot {
-    User(User),
-    Bot(Bot),
-}
-
-#[derive(SimpleObject, Clone)]
-struct User {
-    name: String,
-    email: String,
-}
-
-#[derive(SimpleObject, Clone)]
-struct Bot {
-    id: ID,
-}
-
-#[derive(InputObject)]
-struct BotInput {
-    id: ID,
-}
-
-impl From<&UserOrBot> for UserOrBot {
-    fn from(value: &UserOrBot) -> Self {
-        value.clone()
-    }
-}
-
-#[derive(Debug, InputObject)]
-struct PullRequestsAndIssuesFilters {
-    search: String,
 }
