+++ engine/crates/integration-tests/tests/graphql_connector/defer.rs
@@ -1,11 +1,11 @@
-use integration_tests::{runtime, EngineBuilder, MockGraphQlServer, ResponseExt};
+use integration_tests::{mocks::graphql::FakeGithubSchema, runtime, EngineBuilder, MockGraphQlServer, ResponseExt};
 
 #[test]
 fn test_defer_on_graphql_connector() {
     // Note: this test relies on async-graphql not supporting @defer
     // When that changes we might need to re-think this
     runtime().block_on(async {
-        let graphql_mock = MockGraphQlServer::new().await;
+        let graphql_mock = MockGraphQlServer::new(FakeGithubSchema).await;
 
         let engine = EngineBuilder::new(schema(graphql_mock.port())).build().await;
