+++ engine/crates/integration-tests/tests/graphql_connector/headers.rs
@@ -1,10 +1,10 @@
-use integration_tests::{runtime, EngineBuilder, MockGraphQlServer, ResponseExt};
+use integration_tests::{mocks::graphql::FakeGithubSchema, runtime, EngineBuilder, MockGraphQlServer, ResponseExt};
 use serde_json::json;
 
 #[test]
 fn test_header_forwarding() {
     runtime().block_on(async {
-        let graphql_mock = MockGraphQlServer::new().await;
+        let graphql_mock = MockGraphQlServer::new(FakeGithubSchema).await;
 
         let engine = EngineBuilder::new(schema(graphql_mock.port()))
             .with_env_var("API_KEY", "BLAH")
